+++ packages/ui-core/src/components/Counter/Counter.test.tsx
@@ -1,0 +1,103 @@
+import * as React from 'react';
+import { shallow, mount } from 'enzyme';
+import Counter from './Counter';
+
+const props = {
+    className: 'test-class',
+    initialValue: 10,
+    min: 3,
+    max: 33,
+    isDisabled: true,
+};
+
+describe('<Counter />', () => {
+    describe('layout', () => {
+        it('renders Counter', () => {
+            const wrapper = shallow(
+                <Counter />
+            );
+            expect(wrapper).toMatchSnapshot();
+        });
+
+        it('renders Counter with props', () => {
+            const wrapper = shallow(
+                <Counter {...props} />
+            );
+            expect(wrapper).toMatchSnapshot();
+        });
+    });
+
+    describe('handlers', () => {
+        it('calls onChange for buttons', () => {
+            const handleClick = jest.fn();
+
+            const wrapper = shallow(
+                <Counter initialValue={10} onChange={handleClick} />
+            );
+
+            wrapper.find('button').first().simulate('click');
+            expect(handleClick).toHaveBeenCalledWith(9);
+
+            wrapper.find('button').at(1).simulate('click');
+            expect(handleClick).toHaveBeenCalledWith(10);
+        });
+
+        it('calls onChange for input', () => {
+            const handleChange = jest.fn();
+
+            const wrapper = shallow(
+                <Counter onChange={handleChange} />
+            );
+
+            wrapper.find('input').simulate('change', { target: { value: 22 } });
+            expect(handleChange).toHaveBeenCalledWith(22);
+        });
+
+        it('calls onChange for input with wrong value', () => {
+            const handleChange = jest.fn();
+
+            const wrapper = shallow(
+                <Counter initialValue={10} onChange={handleChange} />
+            );
+
+            wrapper.find('input').simulate('change', { target: { value: 'hello' } });
+            expect(handleChange).not.toBeCalled();
+        });
+
+        it('calls onChange for input with value less than min limit and more than max limit', () => {
+            const handleChange = jest.fn();
+
+            const wrapper = shallow(
+                <Counter min={3} max={33} onChange={handleChange} />
+            );
+
+            wrapper.find('input').simulate('change', { target: { value: 2 } });
+            expect(handleChange).toHaveBeenCalledWith(3);
+
+            wrapper.find('input').simulate('change', { target: { value: 34 } });
+            expect(handleChange).toHaveBeenCalledWith(33);
+        });
+
+        it('calls onChange for left button near min limit', () => {
+            const handleClick = jest.fn();
+
+            const wrapper = mount(
+                <Counter initialValue={3} min={3} onChange={handleClick} />
+            );
+
+            wrapper.find('button').first().simulate('click');
+            expect(handleClick).not.toHaveBeenCalled();
+        });
+
+        it('calls onChange for right button near max limit', () => {
+            const handleClick = jest.fn();
+
+            const wrapper = mount(
+                <Counter initialValue={33} max={33} onChange={handleClick} />
+            );
+
+            wrapper.find('button').at(1).simulate('click');
+            expect(handleClick).not.toHaveBeenCalled();
+        });
+    });
+});