+++ packages/ui-core/src/colors/colorsData.js
@@ -1,0 +1,90 @@
+export default [
+    {
+        title: 'Basic Colors',
+        items: [
+            {
+                name: 'Green',
+                value: '#00B956',
+            },
+            {
+                name: 'Purple',
+                value: '#731982',
+            },
+            {
+                name: 'Clear White',
+                value: '#FFFFFF',
+            },
+            {
+                name: 'Spb Sky 0',
+                value: '#F6F6F6',
+            },
+            {
+                name: 'Spb Sky 1',
+                value: '#EDEDED',
+            },
+            {
+                name: 'Spb Sky 2',
+                value: '#D8D8D8',
+            },
+            {
+                name: 'Fresh Asphalt',
+                value: '#333333',
+            },
+            {
+                name: 'Full Black',
+                value: '#000000',
+            },
+        ],
+    },
+    {
+        title: 'Secondary Colors',
+        items: [
+            {
+                name: 'Warm RedC',
+                value: '#EB5A40',
+            },
+            {
+                name: '137C',
+                value: '#FFA717',
+            },
+
+            {
+                name: '311C',
+                value: '#5BD9E5',
+            },
+            {
+                name: 'Reflex Blue',
+                value: '#444189',
+            },
+        ],
+    },
+    {
+        title: 'System Colors',
+        items: [
+            {
+                name: 'Fury',
+                value: '#F62434',
+            },
+            {
+                name: 'Mandarin',
+                value: '#FFA717',
+            },
+            {
+                name: 'Link',
+                value: '#34AAF2',
+            },
+            {
+                name: 'Button Hover Green',
+                value: '#10E272',
+            },
+            {
+                name: 'Button hover Purple',
+                value: '#A500BF',
+            },
+            {
+                name: 'Button Down',
+                value: '#404D46',
+            },
+        ],
+    },
+];